requirements.txt
opencv-python==4.8.0.74
Pillow==10.0.0
pdf2image==1.16.3
PyPDF2==3.0.1
numpy==1.24.3